refactor: 버스 경로 리스트 아이템 컴포넌트 리팩토링
* 시내버스 시간표에서도 사용 가능하도록 BusRouteListItem.tsx 컴포넌트로 분리
